feat: 특정 점포의 전체 리뷰 조회 구현 - 조회 가능 상태인 리뷰만 내림차순으로 정렬하여 반환

+++ src/main/java/com/dongyang/dongpo/dto/store/ReviewDto.java
@@ -26,7 +26,7 @@ public class ReviewDto {
     private String memberNickname;
     private String memberMainTitle;
     private String memberProfilePic;
-    private Integer reviewStar;
+    private Double reviewStar;
     private String text;
     private List<String> reviewPics;
     private LocalDateTime registerDate;
